Kaushik | Move the gitlab mr plugin to appropriate folder
