+++ .gitignore
@@ -51,10 +51,12 @@ out/
 build
 common/build
 github-pr-status/build
+gitlab-mr-status/build
 gerrit-cs-status/build
 stash-pr-status/build
 common/out
 github-pr-status/out
+gitlab-mr-status/out
 gerrit-cs-status/out
 stash-pr-status/out
