+++ common/src/main/java/com/tw/go/plugin/BuildStatusNotifierPlugin.java
@@ -162,11 +162,8 @@ public abstract class BuildStatusNotifierPlugin implements GoPlugin {
                     String revision = (String) modifications.get(0).get("revision");
                     Map modificationData = (Map) modifications.get(0).get("data");
                     String prId = (String) modificationData.get("PR_ID");
-                    /*
-                    With a standard git backend rather than the github API, PR_ID is null.  Use
-                    current branch identifier in that case.
-                     */
-                    if (prId == null) {
+
+                    if (StringUtils.isEmpty(prId)) {
                         prId = (String) modificationData.get("CURRENT_BRANCH");
                     }
