+++ gitlab-mr-status/build.gradle
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2019 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    id 'java'
+}
+
+group 'com.tw.go.plugin'
+version '1.5'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile project(":common")
+    compile group: 'org.gitlab', name: 'java-gitlab-api', version: '4.1.0'
+
+    testCompile 'org.mockito:mockito-all:1.8.5'
+    testCompile 'junit:junit:4.8.2'
+}
+
+processResources {
+    project.pluginDesc.id = "gitlab.mr.status"
+    project.pluginDesc.name = "GitLab Merge Requests status notifier"
+    project.pluginDesc.description = "Updates build status for GitLab Merge requests"
+
+    from("src/main/resources") {
+        filesMatching('plugin.xml') {
+            expand project.pluginDesc
+        }
+    }
+}
