+++ gitlab-mr-status/src/main/java/com/tw/go/plugin/provider/GitlLabConfigurationView.java
@@ -1,4 +1,4 @@
-package com.tw.go.plugin.provider.gitlab;
+package com.tw.go.plugin.provider;
 
 import com.tw.go.plugin.setting.PluginConfigurationView;
