+++ gitlab-mr-status/src/main/resources/common/plugin-settings-gitlab.template.html
