+++ gitlab-mr-status/src/main/resources/gitlab/defaults.properties
