+++ gitlab-mr-status/src/main/resources/gitlab/plugin.xml
